requirements.txt
Flask==3.0.3
Pillow==10.4.0
pillow-heif==0.18.0
gunicorn==23.0.0